+++ lib/config/content_benchmark.yaml
@@ -41,7 +41,7 @@ extractor:
     config:
       di_class: lib.evagg.llm.openai.OpenAIDotEnvConfig
   mention_finder:
-    di_class: lib.evagg.lit.TruthsetVariantMentionFinder
+    di_class: lib.evagg.content.TruthsetVariantMentionFinder
     entity_annotator: "{{ncbi_client}}"
     gene_lookup_client: "{{ncbi_client}}"
   variant_lookup_client: "{{ncbi_client}}"
