+++ lib/config/tiny_config.yaml
@@ -38,7 +38,7 @@ extractor:
     config:
       di_class: lib.evagg.llm.openai.OpenAIDotEnvConfig
   mention_finder:
-    di_class: lib.evagg.lit.VariantMentionFinder
+    di_class: lib.evagg.content.VariantMentionFinder
     entity_annotator: "{{ncbi_client}}"
     gene_lookup_client: "{{ncbi_client}}"
   variant_lookup_client: "{{ncbi_client}}"
