+++ lib/evagg/__init__.py
@@ -1,9 +1,7 @@
 """The evagg core library."""
 
-from ._content.prompt_based import PromptBasedContentExtractor
-from ._content.simple import SimpleContentExtractor
-from ._content.truth_set import TruthsetContentExtractor
 from .app import SynchronousLocalApp
+from .content import PromptBasedContentExtractor, SimpleContentExtractor, TruthsetContentExtractor
 from .interfaces import IEvAggApp, IExtractFields, IGetPapers, IWriteOutput
 from .io import ConsoleOutputWriter, FileOutputWriter, TableOutputWriter
 from .library import RemoteFileLibrary, SimpleFileLibrary, TruthsetFileLibrary
